Split actions/index into user and auth
