+++ src/containers/app.js
@@ -5,7 +5,8 @@ import { Provider } from 'react-redux'
 import reducer from '../reducers/index'
 import thunkMiddleware from 'redux-thunk'
 import {Scene, Router} from 'react-native-router-flux'
-import { setIsConnected, loadUserData, fetchProjects } from '../actions/index'
+import { setIsConnected, fetchProjects } from '../actions/index'
+import { loadUserData } from '../actions/user'
 
 import ZooniverseApp from './zooniverseApp'
 import ProjectList from '../components/ProjectList'
